Correction + ajout gestion Responsable

+++ crudapi/src/main/java/com/optimizertruck/crudapi/repository/ResponsableRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface ResponsableRepository extends JpaRepository<Responsable,Integer> {
+public interface ResponsableRepository extends JpaRepository<Responsable,Long> {
 }
